#22 Please love me again. Moved generateDefaultButtonNames to own file: gamepadUtils.js, edited imports.
