+++ app/react/components/ButtonNamesManager.js
@@ -1,36 +1,6 @@
 import {useEffect, useState} from 'react';
-import path from 'path';
-import fs from 'fs';
-import userProfile from '../../userProfile';
 import {readButtonNamesFromFile, saveButtonNamesToFile} from '../../local/fileOperations'
-
-const profileDir = path.join(userProfile.getUserDataDir(), 'Frostoven/GamepadProfiler');
-
-const checkAndCreateDir = async () => {
-  try {
-    await fs.promises.access(profileDir);
-  } catch (e) {
-    await fs.promises.mkdir(profileDir, {recursive: true});
-  }
-};
-
-const generateDefaultButtonNames = (gamepad) => {
-  const defaultButtonNames = {};
-
-  if (gamepad && gamepad.buttons) {
-    for (let i = 0; i < gamepad.buttons.length; i++) {
-      defaultButtonNames[`bt${i}`] = `Button ${i}`;
-    }
-  }
-
-  if (gamepad && gamepad.axes) {
-    for (let i = 0; i < gamepad.axes.length; i++) {
-      defaultButtonNames[`ax${i}`] = `Axis ${i}`;
-    }
-  }
-
-  return defaultButtonNames;
-};
+import {generateDefaultButtonNames} from '../../local/gamepadUtils'
 
 const ButtonNamesManager = ({gamepad}) => {
   const [buttonNames, setButtonNames] = useState(() => generateDefaultButtonNames(gamepad));
@@ -60,6 +30,5 @@ const ButtonNamesManager = ({gamepad}) => {
 
 export {
   ButtonNamesManager,
-  checkAndCreateDir,
   generateDefaultButtonNames
 }
